permission system complete

+++ app/Http/Controllers/AuthorContributorController.php
@@ -14,11 +14,13 @@ class AuthorContributorController extends Controller
 
 	public function add_contributor(Request $request)
 	{
+		$this->authorize('new_paper_submission');
 		$this->author_contributor_interface->create($request, session('paper_submission_id'));
 		return back();
 	}
 	public function delete_contributor($id)
 	{
+		$this->authorize('new_paper_submission');
 		$this->author_contributor_interface->delete($id);
 		return back();
 	}
